backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.9.0
opencv-python-headless>=4.8.0
numpy>=1.24.0,<2.0.0
mediapipe>=0.10.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0
gitpython>=3.1.30
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
seaborn>=0.12.0
protobuf>=4.21.0,<5.0.0
transformers>=4.30.0
Pillow>=9.5.0
accelerate
